refactor: Update ExpandPestDetails widget to use bold font for title
The code changes in the ExpandPestDetails widget update the title text to use a bold font style. This change improves the visual presentation of the pest name and enhances the readability of the widget. By using a bold font, the title stands out more prominently and provides a better user experience.