chore: Update user_info_bottomsheet.dart, registration.dart, and sing_up_form.dart
This code change updates the user_info_bottomsheet.dart file by importing the Login page and adding navigation to it when the user logs out. It also updates the registration.dart file by wrapping the SingUpForm widget with an Expanded widget to fix the layout issue. Additionally, it updates the sing_up_form.dart file by adding MainAxisAlignment.center to the column to vertically center the form. These changes improve the functionality and layout of the respective pages.